+++ packages/stream_chat/lib/src/core/models/own_user.dart
@@ -17,6 +17,7 @@ class OwnUser extends User {
     this.unreadChannels = 0,
     this.channelMutes = const [],
     this.unreadThreads = 0,
+    this.blockedUserIds = const [],
     required super.id,
     super.role,
     super.name,
@@ -72,6 +73,7 @@ class OwnUser extends User {
     List<ChannelMute>? channelMutes,
     List<Device>? devices,
     List<Mute>? mutes,
+    List<String>? blockedUserIds,
     int? totalUnreadCount,
     int? unreadChannels,
     int? unreadThreads,
@@ -99,6 +101,7 @@ class OwnUser extends User {
         totalUnreadCount: totalUnreadCount ?? this.totalUnreadCount,
         unreadChannels: unreadChannels ?? this.unreadChannels,
         unreadThreads: unreadThreads ?? this.unreadThreads,
+        blockedUserIds: blockedUserIds ?? this.blockedUserIds,
         language: language ?? this.language,
       );
 
@@ -124,6 +127,7 @@ class OwnUser extends User {
       totalUnreadCount: other.totalUnreadCount,
       unreadChannels: other.unreadChannels,
       unreadThreads: other.unreadThreads,
+      blockedUserIds: other.blockedUserIds,
       updatedAt: other.updatedAt,
       language: other.language,
     );
@@ -153,6 +157,10 @@ class OwnUser extends User {
   @JsonKey(includeIfNull: false)
   final int unreadThreads;
 
+  /// List of user ids that are blocked by the user.
+  @JsonKey(includeIfNull: false)
+  final List<String> blockedUserIds;
+
   /// Known top level fields.
   ///
   /// Useful for [Serializer] methods.
@@ -163,6 +171,7 @@ class OwnUser extends User {
     'unread_channels',
     'channel_mutes',
     'unread_threads',
+    'blocked_user_ids',
     ...User.topLevelFields,
   ];
 }